+++ src/main/java/org/accounting/system/entities/projections/InstallationProjection.java
@@ -1,6 +1,9 @@
 package org.accounting.system.entities.projections;
 
 import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.Setter;
+import org.accounting.system.entities.Capacity;
 import org.accounting.system.entities.MetricDefinition;
 import org.bson.codecs.pojo.annotations.BsonProperty;
 
@@ -30,6 +33,12 @@ public class InstallationProjection {
 
     @BsonProperty("external_id")
     private String externalId;
+
+    @Getter
+    @Setter
+    @BsonProperty("capacities")
+    private List<Capacity> capacities;
+
 
     public String getId() {
         return id;